refactor(downloader): improve .download_data() robustness and file cleanup logic

+++ pydriosm/downloader/_base.py
@@ -238,7 +238,7 @@ class BaseDownloader:
             elif verbose is True or verbose == 1:
                 print("Cancelled.")
 
-            return None
+        return None
 
     @classmethod
     def get_prepacked_data(cls, meth, data_name='<data_name>', file_stem=None, ext=".pkl.xz",
@@ -325,7 +325,8 @@ class BaseDownloader:
                     data = meth(**kwargs)
 
                     if verbose:
-                        indent = len(re.match(r'^ *', ending_message).group())
+                        indent_ = re.match(r'^ *', ending_message)
+                        indent = len(indent_.group()) if indent_ else 0
                         end = "\n  " + " " * indent if verbose == 2 else "\n"
                         print(ending_message, end=end)
 
@@ -355,7 +356,7 @@ class BaseDownloader:
         :param subregion_name: name/URL of a (sub)region available on a free download server
         :type subregion_name: str
         :param valid_names: names of all (sub)regions available on a free download server
-        :type valid_names: typing.Iterable
+        :type valid_names: typing.Collection | None
         :param raise_error: (if the input fails to match a valid name) whether to raise the error
             :py:class:`pydriosm.downloader.InvalidSubregionName`, defaults to ``True``
         :type raise_error: bool
@@ -436,7 +437,7 @@ class BaseDownloader:
             available on a free download server
         :type osm_file_format: str
         :param valid_formats: fil extensions of the data available on a free download server
-        :type valid_formats: typing.Iterable
+        :type valid_formats: typing.Collection | None
         :param raise_error: (if the input fails to match a valid name) whether to raise the error
             :py:class:`pydriosm.downloader.InvalidFileFormatError`, defaults to ``True``
         :type raise_error: bool
@@ -928,26 +929,34 @@ class BaseDownloader:
             if download_dir_ != self.download_dir:
                 self.download_dir = download_dir_
 
-    def _download_data(self, url, path_to_file, interval=0.5, verbose=False, raise_error=False,
-                       print_state="Downloading", pbar_color='green', msg_wrap_limit=None,
-                       verify_download_dir=True, **kwargs):
+    def download_data(self, url, path_to_file, interval=0.5, verbose=False, raise_error=False,
+                      print_state="Downloading", pbar_color='green', msg_wrap_limit=None,
+                      verify_download_dir=True, **kwargs):
         # noinspection PyShadowingNames
         """
-        Download an OSM data file.
+        Download an OSM data file from a URL.
 
-        :param url: a valid URL of an OSM data file
+        :param url: Valid URL of the OSM data file.
         :type url: str
-        :param path_to_file: path where the downloaded OSM data file is saved
+        :param path_to_file: Destination path for the downloaded file.
         :type path_to_file: str
-        :param verbose: whether to print relevant information in console; defaults to ``False``.
+        :param interval: Sleep interval (seconds) after a successful download. Defaults to ``0.5``.
+        :type interval: float | int
+        :param verbose: Whether to print progress; ``2`` for higher verbosity.
+            Defaults to ``False``.
         :type verbose: bool | int
-        :param pbar_color: Custom colour of the progress bar (e.g. 'green', 'yellow');
-            defaults to ``None``.
-        :type pbar_color: str | None
-        :param raise_error: Whether to raise the provided exception;
-            if ``raise_error=False`` (default), the error will be suppressed.
+        :param raise_error: Whether to raise exceptions on failure. Defaults to ``False``.
         :type raise_error: bool
-        :param kwargs: optional parameters of `pyhelpers.ops.download_file_from_url()`_
+        :param print_state: Prefix text for the status message. Defaults to ``"Downloading"``.
+        :type print_state: str
+        :param pbar_color: Color of the progress bar. Defaults to ``'green'``.
+        :type pbar_color: str
+        :param msg_wrap_limit: Maximum character length for printed messages. Defaults to ``None``.
+        :type msg_wrap_limit: int | None
+        :param verify_download_dir: Whether to update the instance's download directory.
+            Defaults to ``True``.
+        :type verify_download_dir: bool
+        :param kwargs: Optional parameters for `pyhelpers.ops.download_file_from_url`_.
 
         .. _`pyhelpers.ops.download_file_from_url()`:
             https://pyhelpers.readthedocs.io/en/latest/_generated/
@@ -966,13 +975,13 @@ class BaseDownloader:
             >>> os.path.exists(path_to_file)
             False
             >>> # Download the PBF data of Rutland
-            >>> bdl._download_data(url, path_to_file, verbose=True)
+            >>> bdl.download_data(url, path_to_file, verbose=True)
             Downloading "rutland-latest.osm.pbf" 100%|██████████| 1.89M/1.89M | 5.64MB/s ...
               Saving "rutland-latest.osm.pbf" to "./tests/osm_data/" ... Done.
             >>> os.path.isfile(path_to_file)
             True
             >>> # Download the data again
-            >>> bdl._download_data(url, path_to_file, verbose=True)
+            >>> bdl.download_data(url, path_to_file, verbose=True)
             Downloading "rutland-latest.osm.pbf" 100%|██████████| 1.89M/1.89M | 4.91MB/s ...
               Updating "rutland-latest.osm.pbf" in "./tests/osm_data/" ... Done.
             >>> os.path.isfile(path_to_file)
@@ -989,17 +998,22 @@ class BaseDownloader:
             Deleting "./tests/osm_data/" ... Done.
         """
 
-        verbose_ = verbose == 2 or False
+        # Normalize verbosity
+        verbose1 = int(verbose) == 1
+        verbose2 = int(verbose) == 2
+
+        # Track if file existed before download to prevent deleting existing data on failure
+        file_existed = os.path.isfile(path_to_file)
 
         _check_saving_path(
-            path=path_to_file, verbose=verbose_, state_verb=print_state, end=" ... ",
+            path=path_to_file, verbose=verbose2, state_verb=print_state, end=" ... ",
             msg_wrap_limit=msg_wrap_limit)
 
         try:
             f = io.StringIO()
             with contextlib.redirect_stdout(f):
                 download_file_from_url(
-                    url=url, path_to_file=path_to_file, verbose=(int(verbose) == 1 or False),
+                    url=url, path_to_file=path_to_file, verbose=verbose1,
                     print_wrap_limit=msg_wrap_limit, pbar_color=pbar_color, **kwargs)
 
             out = f.getvalue()
@@ -1010,16 +1024,17 @@ class BaseDownloader:
                 else:
                     print(out, end="")
 
-            if verbose_:
+            if verbose2:
                 time.sleep(0 if interval is None else interval)
                 print("Done.")
 
         except Exception as e:
-            if os.path.isfile(path_to_file):
+            # Only remove if we created it during this failed attempt
+            if not file_existed and os.path.isfile(path_to_file):
                 os.remove(path_to_file)
 
             _print_failure_message(
-                e, prefix="Failed. Error:", verbose=verbose, raise_error=raise_error)
+                e, prefix="Failed. Error:", verbose=verbose1, raise_error=raise_error)
 
             if not raise_error:
                 return None
